agent-manager/requirements-dev.txt
pytest>=7.0
ruff>=0.12.0
pytest-mock>=3.10
